Routes.py: "/single-item" - place_bid()
Checks loads the bid attempt requested by the user. Compares their attempt with the bid that will actually go through. If the attempts match, allow the update of the attempted bit. If not, deny the bid. 
DBMS level checks that the bid attempt is not a zero or negative number.